+++ tests/testthat/test-geom_bar_text.R
@@ -1,20 +1,26 @@
 context("basic bar plots")
+library(ggplot2)
 
 test_that("simple bar plots and options do not produce errors", {
   expect_silent( {
-    p <- ggplot2::ggplot(altitudes, ggplot2::aes(x = craft, y = altitude, label = altitude)) +
+    p <- ggplot2::ggplot(altitudes, ggplot2::aes(x = craft, y = altitude, 
+                                                 label = altitude)) +
       geom_col() +
       geom_bar_text()
     print(p)
+  } )
 
-    p <- ggplot2::ggplot(coffees, ggplot2::aes(x = coffee, y = proportion, label = ingredient,
-                        fill = ingredient)) +
+  expect_silent( {
+    p <- ggplot2::ggplot(coffees, ggplot2::aes(x = coffee, y = proportion, 
+                                               label = ingredient, fill = ingredient)) +
       geom_col(position = "stack") +
       geom_bar_text(position = "stack", grow = TRUE, reflow = TRUE)
     print(p)
+  } )
 
-    p <- ggplot2::ggplot(coffees, ggplot2::aes(x = coffee, y = proportion, label = ingredient,
-                        fill = ingredient)) +
+  expect_silent( {
+    p <- ggplot2::ggplot(coffees, ggplot2::aes(x = coffee, y = proportion, 
+                                               label = ingredient, fill = ingredient)) +
       geom_col(position = "dodge") +
       geom_bar_text(position = "dodge", grow = TRUE, reflow = TRUE, 
                     place = "left") +